refactor(quickjs): drop the codegen
The last phase 1 item, held back because it was a feature decision rather than
subtraction. ReactUnity never reaches it: nothing under unity/core names a
codegen type, and QuickJSEngine binds through BindingManager.Bind(), the split
phase 1 made for exactly this.

It also could not work here. BindingManager._EmitDelegateMethod built C# with
CodeGenerator and compiled it at runtime as the fallback for a delegate
signature no hand-written template covers -- which is why Codegen/ was reachable
from the reflect-binding path at all -- but the two guards disagreed.
IsCodeEmitSupported() returned true unless NETCOREAPP while Compile was also
compiled out under NET_STANDARD*, and both test projects run
apiCompatibilityLevel 6. So the path ran, built the source, got null from
Compile, dereferenced it, and swallowed the NullReferenceException.

Gone: Binding/Editor/Codegen/ (17 files), the two codegen binding callbacks and
ICodeGenCallback, Generate(TypeBindingFlags) with _WriteCSharp/_WriteTSD and the
codeGenCallback plumbing, _EmitDelegateMethod and the AddAssemblies pair only it
used, the orphaned DocResolver, and UnitylessReflectBind -- the sole caller of
Generate, itself called by nothing. Prefs loses the 23 members left with no
reader and no longer claims to load from js-bridge.json, which nothing has read
for some time; newLineStyle stays because the live newline property is built on
it.

Kept, relocated: CodeGenUtils mixed the runtime compiler in with naming and type
helpers TSTypeNaming and TypeBindingInfo need, so those four are now
BindingUtils. RemoveAt, ToLiteral and ConcatAsLiteral had no readers left.
TextGenerator stays -- it is also BindingManager's log writer.

The cost: a delegate whose signature no template covers can no longer be bound,
in practice ref/out beyond the template set. No behavioural change under Unity,
where that path already returned null; the difference is that
GenerateReflectedDelegateMethod now warns and names the signature instead of
logging a NullReferenceException. Bindings a user already generated keep working
-- they reference runtime types in Binding/, not the generator.

EditMode 346/354, PlayMode 690/701, 0 failed.
