--- unity/quickjs/Runtime/Source/Binding/Editor/BindingCallback/DefaultCodeGenCallback.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2abbf3f5f59520b4aa1f7fe09b245d9f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
