--- unity/quickjs/Runtime/Source/Binding/Editor/BindingCallback/ICodeGenCallback.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 243d60ec4d2b75d46afe43135034ebc8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
