--- unity/quickjs/Runtime/Source/Binding/Editor/BindingCallback/InMemoryCompilationBindingCallback.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a471925b48dd85d488c14e96e8a44506
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
