+++ unity/quickjs/Runtime/Source/Binding/Editor/BindingUtils.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 52ccb1e401018214d89149910b74f52e
