--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 64721f4fe659d7c459237a7bdac240ef
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
