--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_Class.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff17c5a88714df94a97294725ccea4da
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
