--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_DelegateHotfix.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3682fe278e84fc34587bc6416d957f5d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
