--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_DelegateInvoke.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cc2fb77ac8feede4db4ee4b57c262a5f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
