--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_DelegateOperation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e9067139ac30e59468ccfc8599edaa2c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
