--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_DelegateWrapper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9777db33d2236884a9acdb2f29dd7db0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
