--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_Enum.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1f73f8604ea67c24ba833a09112b0129
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
