--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_Event.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5684fc9a5670ec54f8976ee1ff09640f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
