--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_Field.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f5be9d52ad420b54fa791dca9c427d8d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
