--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_Method.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3430939f47b05c34d89895dbcc46bb27
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
