--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_PlainClass.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: da4a004307c395a4f9767d92483529a6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
