--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_PlainMethod.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8ac2ba9dc9657f74b9f5b19a6140c668
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
