--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenHelper_Property.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 46a271593497aa74bb2de85024b04db7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
