--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenUtils.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6ea048735b4a68c48a366a773c8fdb76
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
