--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f027cd41976eb314f947ae8e49d7c264
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
