--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/CodeGenerator_Doc.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 94d186772749d07438f9ad85fd80178a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
