--- unity/quickjs/Runtime/Source/Binding/Editor/Codegen/TSModuleCodeGen.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ccd29405372243b42b9138515a824c03
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
