--- unity/quickjs/Runtime/Source/Binding/Editor/DocResolver.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5b3992bd6a31d204e8265148a767724e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
